This commit was manufactured by cvs2svn to create tag 'release_25_06_2002_grass5_0_0_pre5'.
git-svn-id: https://svn.osgeo.org/grass/grass/tags/release_20020625_grass_5_0_0pre5@7690 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ src/misc/m.kappa/cmd/readin_data.c
@@ -16,16 +16,16 @@ readin_data()
 /* read in data from the input file or stdin */
   if (fscanf (fd, "%d\n", &ncat) != 1)
     G_fatal_error ("ERROR: reading error matrix size");
-  if ((mat = (int **) G_malloc (ncat * sizeof (int *))) == NULL)
+  if ((mat = (long **) G_malloc (ncat * sizeof (long *))) == NULL)
     G_fatal_error ("ERROR: can't allocate memory");
   else
     for (i=0; i<ncat; ++i)
-      if ((mat[i] = (int *) G_malloc (ncat * sizeof (int))) == NULL)
+      if ((mat[i] = (long *) G_malloc (ncat * sizeof (long))) == NULL)
 	G_fatal_error ("ERROR: can't allocate memory");
 
   for (i=0; i<ncat; ++i)
     for (j=0; j<ncat; ++j)
-      if (fscanf (fd, "%d\n", &mat[i][j]) != 1)
+      if (fscanf (fd, "%ld\n", &mat[i][j]) != 1)
 	G_fatal_error ("ERROR: read in error matrix data");
 
   fclose (fd);
